+++ js/fullscreen.js
@@ -1,7 +1,8 @@
 // eslint-disable-next-line no-unused-vars
 function getFullScreenIcon() {
   var content =
-    '<span data-id="fullscreen" class="fullscreen fullscreenicon text-right">' +
+    '<span data-id="fullscreen" class="fullscreen fullscreenicon text-right" ' +
+    'role="button" aria-label="Toggle fullscreen" title="Volledig scherm">' +
     '<em class="fas fa-expand" id="fullScreenToggleIcon" />' +
     '</span>';